refactor after removing multithreading
