+++ src/main/java/com/wine/to/up/winestyle/parser/service/service/KafkaService.java
@@ -1,7 +1,7 @@
 package com.wine.to.up.winestyle.parser.service.service;
 
 public interface KafkaService {
-    void sendAllalcohol();
+    void sendAllAlcohol();
     void sendAllWines();
     void sendAllSparkling();
 }
