+++ src/test/java/com/wine/to/up/winestyle/parser/service/service/implementation/kafka/KafkaSenderServiceTest.java
@@ -1,10 +1,6 @@
 package com.wine.to.up.winestyle.parser.service.service.implementation.kafka;
 
-import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import com.wine.to.up.commonlib.messaging.KafkaMessageSender;
-import com.wine.to.up.parser.common.api.schema.ParserApi;
 import com.wine.to.up.winestyle.parser.service.domain.entity.Alcohol;
-import com.wine.to.up.winestyle.parser.service.service.Director;
 import com.wine.to.up.winestyle.parser.service.service.RepositoryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -15,19 +11,12 @@ import java.util.List;
 import java.util.concurrent.*;
 
 import static org.mockito.Mockito.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 class KafkaSenderServiceTest {
     @Mock
-    Director parserDirector;
-    @Mock
     RepositoryService repositoryService;
     @Mock
-    KafkaMessageSender<ParserApi.WineParsedEvent> kafkaMessageSender;
-    ThreadFactory kafkaSendAllThreadFactory = new ThreadFactoryBuilder()
-            .setNameFormat("Kafka-Sender-%d")
-            .build();
-    ExecutorService kafkaSendAllThreadPool;
+    KafkaSender kafkaSender;
     @InjectMocks
     KafkaSenderService kafkaSenderService;
 
@@ -44,47 +33,34 @@ class KafkaSenderServiceTest {
     void setUp() throws ExecutionException, InterruptedException {
         MockitoAnnotations.initMocks(this);
 
-        ReflectionTestUtils.setField(kafkaSenderService, "parserDirector", parserDirector);
         ReflectionTestUtils.setField(kafkaSenderService, "repositoryService", repositoryService);
-        ReflectionTestUtils.setField(kafkaSenderService, "kafkaMessageSender", kafkaMessageSender);
-        ReflectionTestUtils.setField(kafkaSenderService, "maxThreadCount", 1);
-        ReflectionTestUtils.setField(kafkaSenderService, "timeout", 100);
+        ReflectionTestUtils.setField(kafkaSenderService, "kafkaSender", kafkaSender);
 
         Mockito.when(repositoryService.getAll()).thenReturn(testAlcoholList);
         Mockito.when(repositoryService.getAllWines()).thenReturn(testAlcoholList);
         Mockito.when(repositoryService.getAllSparkling()).thenReturn(testAlcoholList);
+        Mockito.when(kafkaSender.sendAlcoholToKafka(testAlcohol)).thenReturn(1);
 
         Future<Integer> future = mock(Future.class);
         Mockito.when(future.get()).thenReturn(1);
 
-        kafkaSendAllThreadPool = spy(Executors.newFixedThreadPool(2, kafkaSendAllThreadFactory));
-        ReflectionTestUtils.setField(kafkaSenderService, "kafkaSendAllThreadPool", kafkaSendAllThreadPool);
     }
 
     @Test
-    void sendAllalcohol() throws InterruptedException {
-        kafkaSenderService.sendAllalcohol();
-        Mockito.verify(kafkaSendAllThreadPool, times(1)).awaitTermination(100, TimeUnit.MILLISECONDS);
-    }
+    void sendAllAlcohol() {
+        kafkaSenderService.sendAllAlcohol();
+        Mockito.verify(repositoryService, times(1)).getAll();
+        }
 
     @Test
-    void sendAllWines() throws InterruptedException {
+    void sendAllWines() {
         kafkaSenderService.sendAllWines();
-        Mockito.verify(kafkaSendAllThreadPool, times(1)).awaitTermination(100, TimeUnit.MILLISECONDS);
+        Mockito.verify(repositoryService, times(1)).getAllWines();
     }
 
     @Test
-    void sendAllSparkling() throws InterruptedException {
+    void sendAllSparkling() {
         kafkaSenderService.sendAllSparkling();
-        Mockito.verify(kafkaSendAllThreadPool, times(1)).awaitTermination(100, TimeUnit.MILLISECONDS);
-    }
-
-    @Test
-    void initPool() {
-        kafkaSendAllThreadPool.shutdownNow();
-        assertTrue(kafkaSendAllThreadPool.isShutdown());
-        ReflectionTestUtils.invokeMethod(kafkaSenderService, "renewPools");
-        ExecutorService kafkaSendAllThreadPool = (ExecutorService) ReflectionTestUtils.getField(kafkaSenderService, "kafkaSendAllThreadPool");
-        assertFalse(kafkaSendAllThreadPool.isShutdown());
+        Mockito.verify(repositoryService, times(1)).getAllSparkling();
     }
 }
